requirements.txt
apimatic-core==0.2.17
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
asgiref==3.8.1
CacheControl==0.12.14
certifi==2024.12.14
charset-normalizer==3.4.0
deprecation==2.1.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
jsonpickle==3.3.0
jsonpointer==2.4
msgpack==1.1.0
packaging==24.2
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
square==0.0.3
squareup==39.0.0.20241120
urllib3==2.2.3
